Refactor, added styles
